+++ src/models/SettingsModel.php
@@ -13,4 +13,9 @@ class SettingsModel extends Model
      * @var bool Whether the playground should be enabled.
      */
     public $enablePlayground = true;
+
+    /**
+     * @var bool Whether `hx-` attributes should be prefixed with `data-`.
+     */
+    public $hxDataPrefix = false;
 }